+++ src/lichess/mod.rs
@@ -1,0 +1,5 @@
+pub mod models;
+pub mod account;
+pub mod puzzle;
+pub mod game;
+pub mod stream;
