+++ src/state/mod.rs
@@ -3,3 +3,4 @@ pub mod theme_state;
 pub mod game_mode_state;
 pub mod multiplayer_state;
 pub mod ui_state;
+pub mod lichess_state;
