+++ src/ui/components/mod.rs
@@ -1,0 +1,3 @@
+pub mod centered_rect;
+pub mod right_rect;
+pub mod cell;
