+++ src/ui/main_ui.rs
@@ -1,38 +1,51 @@
 use ratatui::{
-    layout::{Constraint, Direction, Layout},
-    prelude::{Alignment, Rect},
-    style::{Color, Modifier, Style, Stylize},
-    text::{Line, Span},
+    prelude::Alignment,
+    style::{Color, Style},
     widgets::{Block, Borders, Paragraph},
     Frame,
 };
 
 use crate::{
-    constants::Popups, game_logic::game::GameState, server::game_server::{get_host_ip, setup_game_server}, ui::popups::{
-        render_credit_popup, render_end_popup, render_enter_game_code_popup,
-        render_enter_lichess_token_popup, render_error_popup, render_help_popup,
-        render_promotion_popup, render_puzzle_end_popup, render_resign_confirmation_popup,
-        render_success_popup,
-    }
+    constants::Popups,
+    server::game_server::{get_host_ip, setup_game_server},
+    ui::{
+        components::centered_rect::centered_rect,
+        game_ui::render_game_ui,
+        menu::{
+            game_mode_menu::render_game_mode_menu, lichess_menu::render_lichess_menu,
+            main_menu::render_menu_ui,
+        },
+        popup::{
+            credits::render_credit_popup,
+            error::render_error_popup,
+            help::render_help_popup,
+            lichess::{
+                game_code::render_enter_game_code_popup, puzzle_end::render_puzzle_end_popup,
+                token::render_enter_lichess_token_popup,
+            },
+            move_input::render_move_input_popup,
+            multiplayer::{
+                enter_ip::render_enter_multiplayer_ip,
+                wait_for_player::render_wait_for_other_player,
+            },
+            pgn::path::render_load_pgn_popup,
+            resignation::render_resign_confirmation_popup,
+            success::render_success_popup,
+        },
+    },
 };
 
-use super::lichess_menu::render_lichess_menu;
 use super::ongoing_games::render_ongoing_games;
 use super::pgn_viewer_ui::render_pgn_viewer;
-use super::popups::{
-    render_enter_multiplayer_ip, render_load_pgn_popup, render_move_input_popup,
-    render_wait_for_other_player,
-};
-use crate::{
-    app::App,
-    constants::{DisplayMode, Pages, TITLE},
-};
+
+use crate::{app::App, constants::Pages};
 use std::path::Path;
 
 /// Renders the user interface widgets.
 pub fn render(app: &mut App, frame: &mut Frame<'_>) {
     let main_area = frame.area();
 
+    //TODO: Use a mega match
     // Solo game
     if app.ui_state.current_page == Pages::Solo {
         render_game_ui(frame, app, main_area);
@@ -76,7 +89,8 @@ pub fn render(app: &mut App, frame: &mut Frame<'_>) {
     else if app.ui_state.current_page == Pages::Bot {
         // Check if engine path is configured
         if app
-            .bot_state.chess_engine_path
+            .bot_state
+            .chess_engine_path
             .as_ref()
             .is_none_or(|path| path.is_empty())
         {
@@ -131,7 +145,7 @@ pub fn render(app: &mut App, frame: &mut Frame<'_>) {
     }
     // Game mode menu
     else if app.ui_state.current_page == Pages::GameModeMenu {
-        crate::ui::game_mode_menu::render_game_mode_menu(frame, app);
+        render_game_mode_menu(frame, app);
     }
     // Ongoing games list
     else if app.ui_state.current_page == Pages::OngoingGames {
@@ -204,7 +218,9 @@ pub fn render(app: &mut App, frame: &mut Frame<'_>) {
             };
 
             // Check if we're still waiting for Elo change calculation
-            let (elo_change, is_calculating) = if let Some(puzzle_game) = &app.lichess_state.puzzle_game {
+            let (elo_change, is_calculating) = if let Some(puzzle_game) =
+                &app.lichess_state.puzzle_game
+            {
                 (
                     puzzle_game.elo_change,
                     puzzle_game.elo_change.is_none() && puzzle_game.elo_change_receiver.is_some(),
@@ -216,472 +232,5 @@ pub fn render(app: &mut App, frame: &mut Frame<'_>) {
             render_puzzle_end_popup(frame, &message, elo_change, is_calculating);
         }
         _ => {}
-    }
-}
-
-/// helper to align pop up to the right of the screen
-pub fn right_rect(percent_x: u16, percent_y: u16, r: Rect) -> Rect {
-    let popup_layout = Layout::default()
-        .direction(Direction::Vertical)
-        .constraints([
-            Constraint::Percentage((100 - percent_y) / 2),
-            Constraint::Percentage(percent_y),
-            Constraint::Percentage((100 - percent_y) / 2),
-        ])
-        .split(r);
-
-    let cols = Layout::default()
-        .direction(Direction::Horizontal)
-        .constraints([
-            Constraint::Percentage(100 - percent_x),
-            Constraint::Percentage(percent_x),
-        ])
-        .split(popup_layout[1]);
-
-    cols[1]
-}
-
-/// Helper function to create a centered rect using up certain percentage of the available rect `r`
-pub fn centered_rect(percent_x: u16, percent_y: u16, r: Rect) -> Rect {
-    let popup_layout = Layout::default()
-        .direction(Direction::Vertical)
-        .constraints([
-            Constraint::Percentage((100 - percent_y) / 2),
-            Constraint::Percentage(percent_y),
-            Constraint::Percentage((100 - percent_y) / 2),
-        ])
-        .split(r);
-
-    Layout::default()
-        .direction(Direction::Horizontal)
-        .constraints([
-            Constraint::Percentage((100 - percent_x) / 2),
-            Constraint::Percentage(percent_x),
-            Constraint::Percentage((100 - percent_x) / 2),
-        ])
-        .split(popup_layout[1])[1]
-}
-
-pub fn render_cell(frame: &mut Frame, square: Rect, color: Color, modifier: Option<Modifier>) {
-    let mut cell = Block::default().bg(color);
-    if let Some(modifier) = modifier {
-        cell = cell.add_modifier(modifier);
-    }
-    frame.render_widget(cell, square);
-}
-
-// Method to render the home menu and the options
-pub fn render_menu_ui(frame: &mut Frame, app: &App, main_area: Rect) {
-    // Determine the "skin" text
-    let display_mode_menu = {
-        let skin_name = match app.game.ui.display_mode {
-            DisplayMode::DEFAULT => "Default",
-            DisplayMode::ASCII => "ASCII",
-            DisplayMode::CUSTOM => app.game.ui.skin.name.as_str(),
-        };
-        format!("Skin: {skin_name}")
-    };
-
-    // Determine the "sound" text (only if sound feature is enabled)
-    #[cfg(feature = "sound")]
-    let sound_menu = {
-        let sound_status = if app.sound_enabled {
-            "On 🔊"
-        } else {
-            "Off 🔇"
-        };
-        format!("Sound: {sound_status}")
-    };
-
-    // Menu items with descriptions
-    let mut menu_items: Vec<(&str, &str)> = vec![
-        ("Play Game", "Local, Multiplayer, or Bot game"),
-        ("Play on Lichess", "Play on Lichess.org"),
-        (&display_mode_menu, "Change display theme"),
-    ];
-
-    // Add sound menu item only if sound feature is enabled
-    #[cfg(feature = "sound")]
-    {
-        menu_items.push((&sound_menu, "Toggle sound effects"));
-    }
-
-    menu_items.extend(vec![
-        ("Help", "View keyboard shortcuts and controls"),
-        ("About", "Project information and credits"),
-    ]);
-
-    // Menu height depends on number of items, each takes 3 lines (item + description/empty + spacing), plus padding
-    let menu_height = menu_items.len() as u16 * 3 + 4;
-
-    let main_layout_horizontal = Layout::default()
-        .direction(Direction::Vertical)
-        .constraints(
-            [
-                Constraint::Ratio(1, 5),         // Title
-                Constraint::Length(1),           // Subtitle
-                Constraint::Min(0),              // Flexible space above menu
-                Constraint::Length(menu_height), // Menu (fixed height)
-                Constraint::Min(0),              // Flexible space below menu
-                Constraint::Ratio(1, 10),        // Footer/hints
-            ]
-            .as_ref(),
-        )
-        .split(main_area);
-
-    // Title
-    let title_paragraph = Paragraph::new(TITLE)
-        .alignment(Alignment::Center)
-        .block(Block::default());
-    frame.render_widget(title_paragraph, main_layout_horizontal[0]);
-
-    // Subtitle
-    let sub_title = Paragraph::new("A chess game made in 🦀")
-        .alignment(Alignment::Center)
-        .block(Block::default());
-    frame.render_widget(sub_title, main_layout_horizontal[1]);
-
-    // Menu area (centered in the middle section)
-    let menu_area = main_layout_horizontal[3];
-
-    // Render menu items
-    let mut menu_lines: Vec<Line<'_>> = vec![];
-
-    for (i, (item, description)) in menu_items.iter().enumerate() {
-        let is_selected = app.ui_state.menu_cursor == i as u8;
-
-        // Create styled menu item
-        let item_style = if is_selected {
-            Style::default()
-                .fg(Color::Black)
-                .bg(Color::LightBlue)
-                .add_modifier(Modifier::BOLD)
-        } else {
-            Style::default().fg(Color::White)
-        };
-
-        // Menu item line with indicator
-        let indicator = if is_selected { "▶ " } else { "  " };
-        let item_text = format!("{}{}", indicator, item);
-        menu_lines.push(Line::from(vec![Span::styled(item_text, item_style)]));
-
-        // Description line (only for selected item to save space)
-        if is_selected {
-            menu_lines.push(Line::from(vec![Span::styled(
-                format!("   {}", description),
-                Style::default()
-                    .fg(Color::Gray)
-                    .add_modifier(Modifier::ITALIC),
-            )]));
-        } else {
-            menu_lines.push(Line::from(""));
-        }
-
-        // Add spacing between menu items
-        menu_lines.push(Line::from(""));
-    }
-
-    let menu_paragraph = Paragraph::new(menu_lines)
-        .alignment(Alignment::Center)
-        .block(Block::default());
-    frame.render_widget(menu_paragraph, menu_area);
-
-    // Footer with keyboard hints
-    let version = env!("CARGO_PKG_VERSION");
-    let footer_text = vec![
-        Line::from(""),
-        Line::from(vec![
-            Span::styled("Navigation: ", Style::default().fg(Color::Gray)),
-            Span::styled("↑/k ", Style::default().fg(Color::Yellow)),
-            Span::styled("↓/j ", Style::default().fg(Color::Yellow)),
-            Span::styled("| Select: ", Style::default().fg(Color::Gray)),
-            Span::styled("Enter/Space", Style::default().fg(Color::Yellow)),
-            Span::styled(" | Help: ", Style::default().fg(Color::Gray)),
-            Span::styled("?", Style::default().fg(Color::Yellow)),
-        ]),
-        Line::from(format!("Version: {}", version))
-            .alignment(Alignment::Center)
-            .style(Style::default().fg(Color::Gray)),
-    ];
-
-    let footer = Paragraph::new(footer_text)
-        .alignment(Alignment::Center)
-        .block(Block::default());
-    frame.render_widget(footer, main_layout_horizontal[5]);
-}
-
-// Method to render the game board and handle game popups
-pub fn render_game_ui(frame: &mut Frame<'_>, app: &mut App, main_area: Rect) {
-    let main_layout_horizontal = Layout::default()
-        .direction(Direction::Vertical)
-        .constraints(
-            [
-                Constraint::Ratio(1, 20),  // Top padding
-                Constraint::Ratio(18, 20), // Board area (increased)
-                Constraint::Min(0),        // Bottom padding (minimal)
-            ]
-            .as_ref(),
-        )
-        .split(main_area);
-
-    let main_layout_vertical = Layout::default()
-        .direction(Direction::Horizontal)
-        .constraints(
-            [
-                Constraint::Ratio(1, 18),  // Left padding (reduced)
-                Constraint::Ratio(1, 18),  // Rank labels (1-8)
-                Constraint::Ratio(11, 18), // Board (increased from 9 to 11)
-                Constraint::Ratio(1, 18),  // Right padding
-                Constraint::Ratio(4, 18),  // Sidebar (reduced from 5 to 4)
-            ]
-            .as_ref(),
-        )
-        .split(main_layout_horizontal[1]);
-
-    // Create layout for board + file labels + clock
-    let has_clock = app.game.logic.clock.is_some();
-    let board_with_labels = if has_clock {
-        Layout::default()
-            .direction(Direction::Vertical)
-            .constraints(
-                [
-                    Constraint::Length(1), // Clock area (small text)
-                    Constraint::Min(0),    // Board (takes remaining space)
-                    Constraint::Length(1), // File labels (A-H) - minimal height
-                ]
-                .as_ref(),
-            )
-            .split(main_layout_vertical[2])
-    } else {
-        Layout::default()
-            .direction(Direction::Vertical)
-            .constraints(
-                [
-                    Constraint::Min(0),    // Board (takes remaining space)
-                    Constraint::Length(1), // File labels (A-H) - minimal height
-                ]
-                .as_ref(),
-            )
-            .split(main_layout_vertical[2])
-    };
-
-    // Render clocks above board if present (small text, no borders)
-    if has_clock {
-        let clock_area = Layout::default()
-            .direction(Direction::Horizontal)
-            .constraints(
-                [
-                    Constraint::Ratio(1, 2), // White clock (left)
-                    Constraint::Ratio(1, 2), // Black clock (right)
-                ]
-                .as_ref(),
-            )
-            .split(board_with_labels[0]);
-
-        use ratatui::style::{Color, Modifier, Style};
-        use ratatui::widgets::Paragraph;
-
-        // White clock (left) - just text
-        if let Some(ref clock) = app.game.logic.clock {
-            let white_time = clock.format_time(shakmaty::Color::White);
-            let is_white_active =
-                clock.is_running && clock.active_color == Some(shakmaty::Color::White);
-            let white_text = format!("White: {}", white_time);
-            let text_width = white_text.len() as u16;
-
-            if is_white_active {
-                // Render background block only for text width
-                let text_area = Rect {
-                    x: clock_area[0].x,
-                    y: clock_area[0].y,
-                    width: text_width.min(clock_area[0].width),
-                    height: clock_area[0].height,
-                };
-                let bg_block = Block::default().style(Style::default().bg(Color::White));
-                frame.render_widget(bg_block, text_area);
-            }
-
-            let white_clock_style = if is_white_active {
-                Style::default()
-                    .fg(Color::Black)
-                    .add_modifier(Modifier::BOLD)
-            } else {
-                Style::default().fg(Color::White)
-            };
-            let white_clock_text = Paragraph::new(white_text)
-                .style(white_clock_style)
-                .alignment(ratatui::layout::Alignment::Left);
-            frame.render_widget(white_clock_text, clock_area[0]);
-        }
-
-        // Black clock (right) - just text
-        if let Some(ref clock) = app.game.logic.clock {
-            let black_time = clock.format_time(shakmaty::Color::Black);
-            let is_black_active =
-                clock.is_running && clock.active_color == Some(shakmaty::Color::Black);
-            let black_text = format!("Black: {}", black_time);
-            let text_width = black_text.len() as u16;
-
-            if is_black_active {
-                // Render background block only for text width, aligned to the right
-                let text_area = Rect {
-                    x: clock_area[1].x + clock_area[1].width.saturating_sub(text_width),
-                    y: clock_area[1].y,
-                    width: text_width.min(clock_area[1].width),
-                    height: clock_area[1].height,
-                };
-                let bg_block = Block::default().style(Style::default().bg(Color::White));
-                frame.render_widget(bg_block, text_area);
-            }
-
-            let black_clock_style = if is_black_active {
-                Style::default()
-                    .fg(Color::Black)
-                    .add_modifier(Modifier::BOLD)
-            } else {
-                Style::default().fg(Color::White)
-            };
-            let black_clock_text = Paragraph::new(black_text)
-                .style(black_clock_style)
-                .alignment(ratatui::layout::Alignment::Right);
-            frame.render_widget(black_clock_text, clock_area[1]);
-        }
-    }
-
-    let board_index = if has_clock { 1 } else { 0 };
-    let file_labels_index = if has_clock { 2 } else { 1 };
-
-    // Split rank label area to match board height
-    let rank_label_area = Layout::default()
-        .direction(Direction::Vertical)
-        .constraints(
-            [
-                Constraint::Min(0),    // Rank labels (aligned with board - takes remaining space)
-                Constraint::Length(1), // Empty space (aligned with file labels - minimal)
-            ]
-            .as_ref(),
-        )
-        .split(main_layout_vertical[1]);
-
-    let right_box_layout = Layout::default()
-        .direction(Direction::Vertical)
-        .constraints(
-            [
-                Constraint::Ratio(2, 15),
-                Constraint::Ratio(11, 15),
-                Constraint::Ratio(2, 15),
-            ]
-            .as_ref(),
-        )
-        .split(main_layout_vertical[4]);
-    // Board block representing the full board div
-    let board_block = Block::default().style(Style::default());
-
-    // We render the board_block in the center layout made above
-    frame.render_widget(board_block.clone(), board_with_labels[board_index]);
-
-    // Split borrows to avoid borrow checker issue
-    let (ui, logic) = (&mut app.game.ui, &app.game.logic);
-
-    // Get the inner area of the board (accounting for any block padding)
-    let board_inner = board_block.inner(board_with_labels[board_index]);
-    ui.board_render(board_inner, frame, logic);
-
-    // Render rank labels (1-8) on the left - aligned with board's inner area
-    ui.render_rank_labels(frame, rank_label_area[0], logic.game_board.is_flipped);
-
-    // Render file labels (A-H) below the board
-    ui.render_file_labels(
-        frame,
-        board_with_labels[file_labels_index],
-        logic.game_board.is_flipped,
-    );
-
-    //top box for black material
-    let black_taken = app.game.logic.game_board.black_taken_pieces();
-    app.game
-        .ui
-        .black_material_render(board_block.inner(right_box_layout[0]), frame, &black_taken);
-
-    // History area
-    app.game
-        .ui
-        .history_render(board_block.inner(right_box_layout[1]), frame, &app.game);
-
-    //bottom box for white material
-    let white_taken = app.game.logic.game_board.white_taken_pieces();
-    let is_puzzle_mode = app.lichess_state.puzzle_game.is_some();
-    if is_puzzle_mode {
-        app.game.ui.white_material_render_with_puzzle_hint(
-            board_block.inner(right_box_layout[2]),
-            frame,
-            &white_taken,
-            true,
-        );
-    } else {
-        app.game.ui.white_material_render(
-            board_block.inner(right_box_layout[2]),
-            frame,
-            &white_taken,
-        );
-    }
-
-    if app.game.logic.game_state == GameState::Promotion {
-        render_promotion_popup(frame, app);
-    }
-
-    // If the game ended (checkmate or draw) and there's no active popup yet,
-    // open the EndScreen popup so it appears immediately instead of waiting for
-    // another user interaction.
-    if app.game.logic.game_state == GameState::Checkmate {
-        // When game ended by time, player_turn is already set to the winner
-        // For checkmate, the winner is the other player (the one who delivered checkmate)
-        let victorious_player = if app.game.logic.game_ended_by_time {
-            app.game.logic.player_turn
-        } else {
-            app.game.logic.player_turn.other()
-        };
-
-        let string_color = match victorious_player {
-            shakmaty::Color::White => "White",
-            shakmaty::Color::Black => "Black",
-        };
-
-        if app.ui_state.current_popup == Some(Popups::EndScreen) {
-            // Check if it's Lichess multiplayer (restart not available in Lichess)
-            let is_lichess = app
-                .game
-                .logic
-                .opponent
-                .as_ref()
-                .map(|opp| opp.is_lichess())
-                .unwrap_or(false);
-
-            // Check if game ended due to time
-            let message = if app.game.logic.game_ended_by_time {
-                let time_up_color = match victorious_player {
-                    shakmaty::Color::White => "Black",
-                    shakmaty::Color::Black => "White",
-                };
-                format!("{string_color} Won \n{time_up_color} ran out of time")
-            } else {
-                format!("{string_color} Won")
-            };
-
-            render_end_popup(frame, &message, is_lichess);
-        }
-    }
-
-    if app.game.logic.game_state == GameState::Draw && app.ui_state.current_popup == Some(Popups::EndScreen)
-    {
-        // Check if it's Lichess multiplayer (restart not available in Lichess)
-        let is_lichess = app
-            .game
-            .logic
-            .opponent
-            .as_ref()
-            .map(|opp| opp.is_lichess())
-            .unwrap_or(false);
-        render_end_popup(frame, "That's a draw", is_lichess);
     }
 }
