+++ src/ui/menu/mod.rs
@@ -1,0 +1,3 @@
+pub mod game_mode_menu;
+pub mod lichess_menu;
+pub mod main_menu;
