+++ src/ui/mod.rs
@@ -1,8 +1,9 @@
-pub mod game_mode_menu;
-pub mod lichess_menu;
 pub mod main_ui;
 pub mod ongoing_games;
 pub mod pgn_viewer_ui;
-pub mod popups;
 pub mod prompt;
 pub mod tui;
+pub mod popup;
+pub mod menu;
+pub mod components;
+pub mod game_ui;
