+++ src/ui/popup/lichess/mod.rs
@@ -1,0 +1,3 @@
+pub mod game_code;
+pub mod puzzle_end;
+pub mod token;
