+++ src/ui/popup/mod.rs
@@ -1,0 +1,11 @@
+pub mod credits;
+pub mod end;
+pub mod error;
+pub mod help;
+pub mod lichess;
+pub mod move_input;
+pub mod multiplayer;
+pub mod pgn;
+pub mod promotion;
+pub mod resignation;
+pub mod success;
