+++ src/ui/popup/pgn/mod.rs
@@ -1,0 +1,2 @@
+pub mod help;
+pub mod path;
