--- src/ui/popups.rs
@@ -1,847 +1,0 @@
-use std::net::IpAddr;
-
-use super::prompt::Prompt;
-use crate::ui::main_ui::right_rect;
-use crate::{
-    app::App,
-    constants::{NETWORK_PORT, WHITE},
-    pieces::{bishop::Bishop, knight::Knight, queen::Queen, rook::Rook, PieceSize},
-    ui::main_ui::centered_rect,
-};
-use ratatui::{
-    layout::{Alignment, Constraint, Direction, Layout, Position},
-    style::{Color, Modifier, Style, Stylize},
-    text::{Line, Span},
-    widgets::{Block, BorderType, Borders, Clear, Padding, Paragraph, Wrap},
-    Frame,
-};
-
-// This renders a confirmation popup for resigning a game
-pub fn render_resign_confirmation_popup(frame: &mut Frame, app: &App) {
-    let block = Block::default()
-        .title("Resign Game")
-        .borders(Borders::ALL)
-        .border_type(BorderType::Rounded)
-        .padding(Padding::horizontal(1))
-        .border_style(Style::default().fg(WHITE));
-    let area = centered_rect(50, 30, frame.area());
-
-    let opponent_name = if let Some(game) = app.lichess_state.ongoing_games.get(app.ui_state.menu_cursor as usize) {
-        format!("vs {}", game.opponent.username)
-    } else {
-        "this game".to_string()
-    };
-
-    let text = vec![
-        Line::from(""),
-        Line::from(format!(
-            "Are you sure you want to resign {}?",
-            opponent_name
-        ))
-        .alignment(Alignment::Center),
-        Line::from(""),
-        Line::from("This action cannot be undone.").alignment(Alignment::Center),
-        Line::from(""),
-        Line::from(""),
-        Line::from(vec![
-            Span::styled(
-                "Y",
-                Style::default()
-                    .fg(Color::Green)
-                    .add_modifier(Modifier::BOLD),
-            ),
-            Span::raw("es / "),
-            Span::styled(
-                "N",
-                Style::default().fg(Color::Red).add_modifier(Modifier::BOLD),
-            ),
-            Span::raw("o"),
-        ])
-        .alignment(Alignment::Center),
-    ];
-
-    let paragraph = Paragraph::new(text)
-        .block(block.clone())
-        .alignment(Alignment::Left)
-        .wrap(Wrap { trim: true });
-
-    frame.render_widget(Clear, area); //this clears out the background
-    frame.render_widget(block, area);
-    frame.render_widget(paragraph, area);
-}
-
-// This renders a generic error popup with a custom message
-pub fn render_error_popup(frame: &mut Frame, error_message: &str) {
-    let block = Block::default()
-        .title("Error")
-        .borders(Borders::ALL)
-        .border_type(BorderType::Rounded)
-        .padding(Padding::horizontal(1))
-        .border_style(Style::default().fg(Color::Red));
-    let area = centered_rect(50, 30, frame.area());
-
-    // Split the error message by newlines and create a Line for each part
-    let mut text = vec![Line::from("")];
-
-    for line in error_message.split('\n') {
-        if !line.is_empty() {
-            text.push(Line::from(line).alignment(Alignment::Center));
-        } else {
-            text.push(Line::from(""));
-        }
-    }
-
-    text.extend(vec![
-        Line::from(""),
-        Line::from("Press `Esc` or `Enter` to close.").alignment(Alignment::Center),
-    ]);
-
-    let paragraph = Paragraph::new(text)
-        .block(block.clone())
-        .alignment(Alignment::Left)
-        .wrap(Wrap { trim: true });
-
-    frame.render_widget(Clear, area); //this clears out the background
-    frame.render_widget(block, area);
-    frame.render_widget(paragraph, area);
-}
-
-// This renders a generic success popup with a custom message
-pub fn render_success_popup(frame: &mut Frame, success_message: &str) {
-    let block = Block::default()
-        .title("Success")
-        .borders(Borders::ALL)
-        .border_type(BorderType::Rounded)
-        .padding(Padding::horizontal(1))
-        .border_style(Style::default().fg(Color::Green));
-    let area = centered_rect(50, 30, frame.area());
-
-    // Split the success message by newlines and create a Line for each part
-    let mut text = vec![Line::from("")];
-
-    for line in success_message.split('\n') {
-        if !line.is_empty() {
-            text.push(Line::from(line).alignment(Alignment::Center));
-        } else {
-            text.push(Line::from(""));
-        }
-    }
-
-    text.extend(vec![
-        Line::from(""),
-        Line::from("Press `Esc` or `Enter` to close.").alignment(Alignment::Center),
-    ]);
-
-    let paragraph = Paragraph::new(text)
-        .block(block.clone())
-        .alignment(Alignment::Left)
-        .wrap(Wrap { trim: true });
-
-    frame.render_widget(Clear, area); //this clears out the background
-    frame.render_widget(block, area);
-    frame.render_widget(paragraph, area);
-}
-
-// This renders a popup for a promotion
-pub fn render_end_popup(frame: &mut Frame, sentence: &str, is_lichess: bool) {
-    let block = Block::default()
-        .title("Game Over")
-        .title_alignment(Alignment::Center)
-        .borders(Borders::ALL)
-        .border_type(BorderType::Double)
-        .padding(Padding::horizontal(2))
-        .border_style(Style::default().fg(Color::Yellow))
-        .style(Style::default().bg(Color::DarkGray));
-    let area = centered_rect(50, 50, frame.area());
-
-    // Create styled text with better formatting
-    let mut text = vec![Line::from(""), Line::from("")];
-
-    // Split sentence by newlines to handle multi-line messages
-    for line in sentence.lines() {
-        text.push(
-            Line::from(line).alignment(Alignment::Center).style(
-                Style::default()
-                    .fg(Color::Yellow)
-                    .add_modifier(Modifier::BOLD),
-            ),
-        );
-    }
-
-    text.extend(vec![
-        Line::from(""),
-        Line::from(""),
-        Line::from("━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━")
-            .style(Style::default().fg(Color::Gray)),
-        Line::from(""),
-        Line::from("Press `H` to hide this screen")
-            .alignment(Alignment::Center)
-            .style(Style::default().fg(Color::LightBlue)),
-    ]);
-
-    // Only show restart option for non-Lichess games (Lichess games can't be restarted)
-    if !is_lichess {
-        text.push(Line::from(""));
-        text.push(
-            Line::from("Press `R` to restart a new game")
-                .alignment(Alignment::Center)
-                .style(Style::default().fg(Color::LightGreen)),
-        );
-    }
-
-    text.push(Line::from(""));
-    text.push(
-        Line::from("Press `B` to go back to the menu")
-            .alignment(Alignment::Center)
-            .style(Style::default().fg(Color::LightCyan)),
-    );
-    text.push(Line::from(""));
-    text.push(Line::from(""));
-
-    let paragraph = Paragraph::new(text)
-        .block(block.clone())
-        .alignment(Alignment::Center)
-        .wrap(Wrap { trim: true });
-
-    frame.render_widget(Clear, area); //this clears out the background
-    frame.render_widget(block, area);
-    frame.render_widget(paragraph, area);
-}
-
-// This renders a popup for puzzle completion
-pub fn render_puzzle_end_popup(
-    frame: &mut Frame,
-    sentence: &str,
-    elo_change: Option<i32>,
-    is_calculating: bool,
-) {
-    let block = Block::default()
-        .title("Puzzle Complete")
-        .title_alignment(Alignment::Center)
-        .borders(Borders::ALL)
-        .border_type(BorderType::Double)
-        .padding(Padding::horizontal(2))
-        .border_style(Style::default().fg(Color::Yellow))
-        .style(Style::default().bg(Color::DarkGray));
-    let area = centered_rect(50, 50, frame.area());
-
-    // Create styled text with better formatting
-    let mut text = vec![
-        Line::from(""),
-        Line::from(""),
-        Line::from(sentence).alignment(Alignment::Center).style(
-            Style::default()
-                .fg(Color::Yellow)
-                .add_modifier(Modifier::BOLD),
-        ),
-    ];
-
-    // Add Elo change if available, or show calculating message
-    if let Some(change) = elo_change {
-        text.push(Line::from(""));
-        let (change_text, color) = if change > 0 {
-            (format!("+{} Elo", change), Color::Green)
-        } else if change < 0 {
-            (format!("{} Elo", change), Color::Red)
-        } else {
-            ("+0 Elo".to_string(), Color::Yellow)
-        };
-        text.push(
-            Line::from(change_text)
-                .alignment(Alignment::Center)
-                .style(Style::default().fg(color).add_modifier(Modifier::BOLD)),
-        );
-    } else if is_calculating {
-        text.push(Line::from(""));
-        text.push(
-            Line::from("Calculating Elo change...")
-                .alignment(Alignment::Center)
-                .style(Style::default().fg(Color::Cyan)),
-        );
-    }
-
-    text.extend(vec![
-        Line::from(""),
-        Line::from(""),
-        Line::from("━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━")
-            .style(Style::default().fg(Color::Gray)),
-        Line::from(""),
-        Line::from("Press `H` or `Esc` to hide this screen")
-            .alignment(Alignment::Center)
-            .style(Style::default().fg(Color::LightBlue)),
-        Line::from(""),
-        Line::from("Press `N` for a new puzzle")
-            .alignment(Alignment::Center)
-            .style(Style::default().fg(Color::LightGreen)),
-        Line::from(""),
-        Line::from("Press `B` to go back to the menu")
-            .alignment(Alignment::Center)
-            .style(Style::default().fg(Color::LightCyan)),
-        Line::from(""),
-        Line::from(""),
-    ]);
-
-    let paragraph = Paragraph::new(text)
-        .block(block.clone())
-        .alignment(Alignment::Center)
-        .wrap(Wrap { trim: true });
-
-    frame.render_widget(Clear, area); //this clears out the background
-    frame.render_widget(block, area);
-    frame.render_widget(paragraph, area);
-}
-
-// This renders a popup for a promotion
-pub fn render_promotion_popup(frame: &mut Frame, app: &mut App) {
-    let block = Block::default()
-        .title("Pawn promotion")
-        .borders(Borders::ALL)
-        .border_type(BorderType::Rounded)
-        .padding(Padding::horizontal(1))
-        .border_style(Style::default().fg(WHITE));
-    let area = centered_rect(40, 40, frame.area());
-
-    let text = vec![
-        Line::from(""),
-        Line::from("-- Choose your pawn promotion --").alignment(Alignment::Center),
-        Line::from(""),
-    ];
-
-    let paragraph = Paragraph::new(text)
-        .block(Block::default())
-        .alignment(Alignment::Left)
-        .wrap(Wrap { trim: true });
-    frame.render_widget(Clear, area); //this clears out the background
-    frame.render_widget(block, area);
-    frame.render_widget(paragraph, area);
-
-    let inner_popup_layout_vertical = Layout::default()
-        .direction(Direction::Vertical)
-        .constraints(
-            [
-                Constraint::Ratio(1, 3),
-                Constraint::Ratio(1, 3),
-                Constraint::Ratio(1, 3),
-            ]
-            .as_ref(),
-        )
-        .split(area);
-
-    let inner_popup_layout_horizontal = Layout::default()
-        .direction(Direction::Horizontal)
-        .constraints(
-            [
-                Constraint::Ratio(1, 4),
-                Constraint::Ratio(1, 4),
-                Constraint::Ratio(1, 4),
-                Constraint::Ratio(1, 4),
-            ]
-            .as_ref(),
-        )
-        .split(inner_popup_layout_vertical[1]);
-
-    // When a promotion is happening, the mouse should be able to know where the icons are
-    app.game.ui.top_x = inner_popup_layout_horizontal[0].x;
-    app.game.ui.top_y = inner_popup_layout_horizontal[0].y;
-    app.game.ui.width = inner_popup_layout_horizontal[0].width;
-    app.game.ui.height = inner_popup_layout_horizontal[0].height;
-
-    let piece_style = &app.game.ui.skin.piece_style;
-    let piece_size = PieceSize::from_dimensions(inner_popup_layout_horizontal[0].height);
-    let piece_color = Some(app.game.logic.player_turn);
-
-    let piece_styles = &app.game.ui.available_piece_styles;
-    let queen_p = Paragraph::new(Queen::to_string(
-        piece_style.clone(),
-        piece_size,
-        piece_color,
-        piece_styles,
-    ))
-    .block(Block::default())
-    .alignment(Alignment::Center)
-    .style(Style::default().bg(if app.game.ui.promotion_cursor == 0 {
-        Color::LightBlue
-    } else {
-        Color::Reset // Set to the default background color when the condition is false
-    }));
-    frame.render_widget(queen_p, inner_popup_layout_horizontal[0]);
-    let rook_p = Paragraph::new(Rook::to_string(
-        piece_style.clone(),
-        piece_size,
-        piece_color,
-        piece_styles,
-    ))
-    .block(Block::default())
-    .alignment(Alignment::Center)
-    .style(Style::default().bg(if app.game.ui.promotion_cursor == 1 {
-        Color::LightBlue
-    } else {
-        Color::Reset // Set to the default background color when the condition is false
-    }));
-    frame.render_widget(rook_p, inner_popup_layout_horizontal[1]);
-    let bishop_p = Paragraph::new(Bishop::to_string(
-        piece_style.clone(),
-        piece_size,
-        piece_color,
-        piece_styles,
-    ))
-    .block(Block::default())
-    .alignment(Alignment::Center)
-    .style(Style::default().bg(if app.game.ui.promotion_cursor == 2 {
-        Color::LightBlue
-    } else {
-        Color::Reset // Set to the default background color when the condition is false
-    }));
-    frame.render_widget(bishop_p, inner_popup_layout_horizontal[2]);
-    let knight_p = Paragraph::new(Knight::to_string(
-        piece_style.clone(),
-        piece_size,
-        piece_color,
-        piece_styles,
-    ))
-    .block(Block::default())
-    .alignment(Alignment::Center)
-    .style(Style::default().bg(if app.game.ui.promotion_cursor == 3 {
-        Color::LightBlue
-    } else {
-        Color::Reset // Set to the default background color when the condition is false
-    }));
-    frame.render_widget(knight_p, inner_popup_layout_horizontal[3]);
-}
-
-// This render the credit popup
-pub fn render_credit_popup(frame: &mut Frame) {
-    let block = Block::default()
-        .title("Credits")
-        .borders(Borders::ALL)
-        .border_type(BorderType::Rounded)
-        .padding(Padding::horizontal(1))
-        .border_style(Style::default().fg(WHITE));
-    let area = centered_rect(40, 40, frame.area());
-
-    let credits_text = vec![
-        Line::from(""),
-        Line::from("Hi 👋, I'm Thomas, a 23 years old French computer engineer."),
-        Line::from("Thank you for playing Chess-tui! This project started as a personal journey to improve my algorithmic skills and learn Rust."),
-        Line::from(""),
-        Line::from("The entire source code is available on GitHub at https://github.com/thomas-mauran/chess-tui"),
-        Line::from(""),
-        Line::from("Feel free to contribute by picking an issue or creating a new one and star the repo if you wanna support the project !"),
-        Line::from(""),
-        Line::from("Special thanks to my classmates for their support and inspiration!"),
-        Line::from(""),
-        Line::from(""),
-        Line::from(""),
-        Line::from(""),
-        Line::from("Press `Esc` to close the popup.").alignment(Alignment::Center),
-    ];
-
-    // Assuming Alignment is an enum or struct you have defined
-
-    let paragraph = Paragraph::new(credits_text)
-        .block(block.clone())
-        .alignment(Alignment::Left)
-        .wrap(Wrap { trim: true });
-
-    frame.render_widget(Clear, area); //this clears out the background
-    frame.render_widget(block, area);
-    frame.render_widget(paragraph, area);
-}
-
-/// Help popup shown while the PGN viewer is active. Covers only the controls
-/// that make sense for a read-only replay - no piece-selection or skin cycling.
-fn render_pgn_help_popup(frame: &mut Frame) {
-    let block = Block::default()
-        .title("PGN viewer - controls")
-        .borders(Borders::ALL)
-        .border_type(BorderType::Rounded)
-        .padding(Padding::horizontal(1))
-        .border_style(Style::default().fg(WHITE));
-    let area = centered_rect(45, 55, frame.area());
-
-    let text = vec![
-        Line::from("Navigation".underlined().bold()),
-        Line::from(""),
-        Line::from("→ / N / l : Next move"),
-        Line::from("← / P / h : Previous move"),
-        Line::from("g         : Jump to start"),
-        Line::from("G         : Jump to end"),
-        Line::from("Tab       : Next game (multi-game PGN)"),
-        Line::from(""),
-        Line::from("Auto-play".underlined().bold()),
-        Line::from(""),
-        Line::from("Space     : Play / Pause"),
-        Line::from("+  /  -   : Speed up / slow down"),
-        Line::from("            (0.5x, 1x, 1.5x, 2x, 2.5x, 3x, 4x)"),
-        Line::from(""),
-        Line::from("Other".underlined().bold()),
-        Line::from(""),
-        Line::from("h         : Hide the end-of-game banner"),
-        Line::from("?         : Toggle this help"),
-        Line::from("Esc / q   : Exit the viewer"),
-        Line::from(""),
-        Line::from("Press `Esc` or `?` to close.").alignment(Alignment::Center),
-    ];
-
-    let paragraph = Paragraph::new(text)
-        .block(block.clone())
-        .alignment(Alignment::Left)
-        .wrap(Wrap { trim: true });
-
-    frame.render_widget(Clear, area);
-    frame.render_widget(block, area);
-    frame.render_widget(paragraph, area);
-}
-
-// This render the help popup
-pub fn render_help_popup(frame: &mut Frame, app: &crate::app::App) {
-    if app.ui_state.current_page == crate::constants::Pages::PgnViewer {
-        render_pgn_help_popup(frame);
-        return;
-    }
-
-    let block = Block::default()
-        .title("Help menu")
-        .borders(Borders::ALL)
-        .border_type(BorderType::Rounded)
-        .padding(Padding::horizontal(1))
-        .border_style(Style::default().fg(WHITE));
-    let area = centered_rect(40, 65, frame.area());
-
-    // Check if we're playing against a bot (history navigation only in solo mode)
-    let is_solo_mode = app.game.logic.bot.is_none() && app.game.logic.opponent.is_none();
-    let is_puzzle_mode = app.lichess_state.puzzle_game.is_some();
-
-    let mut text = vec![
-        Line::from("Game controls:".underlined().bold()),
-        Line::from(""),
-    ];
-
-    // In puzzle mode, 'h' is for hint, not left movement
-    if is_puzzle_mode {
-        text.push(Line::from(vec![
-            "← ↑/k ↓/j →/l: Use these keys or the mouse to move the ".into(),
-            "blue".blue(),
-            " cursor".into(),
-        ]));
-        text.push(Line::from(""));
-        text.push(Line::from(
-            "T: Show hint (select the piece to move)".yellow(),
-        ));
-        text.push(Line::from(""));
-    } else {
-        text.push(Line::from(vec![
-            "←/h ↑/k ↓/j →/l: Use these keys or the mouse to move the ".into(),
-            "blue".blue(),
-            " cursor".into(),
-        ]));
-        text.push(Line::from(""));
-    }
-
-    text.extend(vec![
-        Line::from("`Ctrl` '+' or '-': Zoom in or out to adjust pieces sizes"),
-        Line::from("(Might differ in certain terminals)"),
-        Line::from(""),
-        Line::from("`Space`: Select a piece"),
-        Line::from(""),
-        Line::from("`Esc`: Deselect a piece / hide popups"),
-        Line::from(""),
-        Line::from("q: Quit the game"),
-        Line::from(""),
-        Line::from("b: Go to the home menu / reset the game"),
-        Line::from(""),
-        Line::from("s: Cycle through available skins"),
-        Line::from(""),
-    ]);
-
-    // Only show history navigation controls in solo mode (not against bot or puzzle)
-    if is_solo_mode && !is_puzzle_mode {
-        text.push(Line::from("P: Navigate to previous position in history"));
-        text.push(Line::from(""));
-        text.push(Line::from("N: Navigate to next position in history"));
-        text.push(Line::from(""));
-    }
-
-    text.extend(vec![
-        Line::from(""),
-        Line::from("Color codes:".underlined().bold()),
-        Line::from(""),
-        Line::from(vec!["Blue cell".blue(), ": Your cursor ".into()]),
-        Line::from(""),
-        Line::from(vec!["Green cell".green(), ": Selected Piece ".into()]),
-        Line::from(""),
-        Line::from(vec![
-            "Purple cell".magenta(),
-            ": The king is getting checked ".into(),
-        ]),
-        Line::from(""),
-        Line::from("Grey cell: Available cells for the selected piece"),
-        Line::from(""),
-        Line::from(""),
-        Line::from("Press `Esc` to close the popup.").alignment(Alignment::Center),
-    ]);
-
-    let text = text;
-
-    let paragraph = Paragraph::new(text)
-        .block(block.clone())
-        .alignment(Alignment::Left)
-        .wrap(Wrap { trim: true });
-
-    frame.render_widget(Clear, area); //this clears out the background
-    frame.render_widget(block, area);
-    frame.render_widget(paragraph, area);
-}
-
-// MULTIPLAYER POPUPS
-// This renders a popup indicating we are waiting for the other player
-pub fn render_wait_for_other_player(frame: &mut Frame, ip: Option<IpAddr>) {
-    let block = Block::default()
-        .title("Waiting ...")
-        .borders(Borders::ALL)
-        .border_type(BorderType::Rounded)
-        .padding(Padding::horizontal(1))
-        .border_style(Style::default().fg(WHITE));
-    let area = centered_rect(40, 40, frame.area());
-
-    let ip_str = ip
-        .map(|i| i.to_string())
-        .unwrap_or_else(|| "Unknown".to_string());
-
-    let text = vec![
-        Line::from(""),
-        Line::from(""),
-        Line::from("Waiting for other player").alignment(Alignment::Center),
-        Line::from(format!(
-            "Host IP address and port: {}:{}",
-            ip_str, NETWORK_PORT
-        ))
-        .alignment(Alignment::Center),
-    ];
-
-    let paragraph = Paragraph::new(text)
-        .block(block.clone())
-        .alignment(Alignment::Left)
-        .wrap(Wrap { trim: true });
-
-    frame.render_widget(Clear, area); //this clears out the background
-    frame.render_widget(block, area);
-    frame.render_widget(paragraph, area);
-}
-
-// allow user to input moves by input text
-pub fn render_move_input_popup(frame: &mut Frame, prompt: &Prompt) {
-    let block = Block::default()
-        .title("Input a move")
-        .borders(Borders::ALL)
-        .border_type(BorderType::Rounded)
-        .padding(Padding::horizontal(1))
-        .border_style(Style::default().fg(WHITE));
-    let area = right_rect(23, 40, frame.area());
-
-    let current_input = prompt.input.as_str();
-    let text = vec![
-        Line::from("Enter a move in chess notation").alignment(Alignment::Center),
-        Line::from(""),
-        Line::from(current_input),
-        Line::from(""),
-        Line::from(
-            "Visit this website for more information: https://www.chess.com/terms/chess-notation",
-        ),
-        Line::from(""),
-        Line::from("Press `Esc` to close the popup.").alignment(Alignment::Center),
-    ];
-
-    let paragraph = Paragraph::new(text)
-        .block(block.clone())
-        .alignment(Alignment::Left)
-        .wrap(Wrap { trim: true });
-
-    frame.set_cursor_position(Position::new(
-        area.x + prompt.character_index as u16 + 2,
-        area.y + 3,
-    ));
-
-    frame.render_widget(Clear, area);
-    frame.render_widget(block, area);
-    frame.render_widget(paragraph, area);
-}
-
-// This renders a popup allowing us to get a user input
-pub fn render_enter_multiplayer_ip(frame: &mut Frame, prompt: &Prompt) {
-    let block = Block::default()
-        .title("Join a game")
-        .borders(Borders::ALL)
-        .border_type(BorderType::Rounded)
-        .padding(Padding::horizontal(1))
-        .border_style(Style::default().fg(WHITE));
-    let area = centered_rect(40, 40, frame.area());
-
-    let current_input = prompt.input.as_str();
-
-    let text = vec![
-        Line::from("Enter the ip address and port of the host:").alignment(Alignment::Center),
-        Line::from(""),
-        Line::from(current_input),
-        Line::from(""),
-        Line::from(""),
-        Line::from(""),
-        Line::from(""),
-        Line::from(format!("Example: 10.111.6.50:{};", NETWORK_PORT)),
-        Line::from("Documentation: https://thomas-mauran.github.io/chess-tui/docs/Multiplayer/Online%20multiplayer/"),
-        Line::from(""),
-        Line::from(""),
-        Line::from("Press `Esc` to close the popup.").alignment(Alignment::Center),
-    ];
-
-    let paragraph = Paragraph::new(text)
-        .block(block.clone())
-        .alignment(Alignment::Left)
-        .wrap(Wrap { trim: true });
-
-    frame.set_cursor_position(Position::new(
-        // Draw the cursor at the current position in the input field.
-        // This position is can be controlled via the left and right arrow key
-        area.x + prompt.character_index as u16 + 2,
-        // Move one line down, from the border to the input line
-        area.y + 3,
-    ));
-
-    frame.render_widget(Clear, area); //this clears out the background
-    frame.render_widget(block, area);
-    frame.render_widget(paragraph, area);
-}
-
-// This renders a popup for entering the path to a PGN file
-pub fn render_load_pgn_popup(frame: &mut Frame, prompt: &Prompt) {
-    let block = Block::default()
-        .title("Load PGN")
-        .borders(Borders::ALL)
-        .border_type(BorderType::Rounded)
-        .padding(Padding::horizontal(1))
-        .border_style(Style::default().fg(WHITE));
-    let area = centered_rect(60, 35, frame.area());
-
-    let current_input = prompt.input.as_str();
-
-    let text = vec![
-        Line::from("Enter the absolute path to a .pgn file:").alignment(Alignment::Center),
-        Line::from(""),
-        Line::from(current_input),
-        Line::from(""),
-        Line::from(""),
-        Line::from("Multi-game PGN files are supported."),
-        Line::from(""),
-        Line::from("Press `Enter` to load, `Esc` to cancel.").alignment(Alignment::Center),
-    ];
-
-    let paragraph = Paragraph::new(text)
-        .block(block.clone())
-        .alignment(Alignment::Left)
-        .wrap(Wrap { trim: true });
-
-    frame.set_cursor_position(Position::new(
-        area.x + prompt.character_index as u16 + 2,
-        area.y + 3,
-    ));
-
-    frame.render_widget(Clear, area);
-    frame.render_widget(block, area);
-    frame.render_widget(paragraph, area);
-}
-
-// This renders a popup allowing us to enter a game code to join a Lichess game
-pub fn render_enter_game_code_popup(frame: &mut Frame, prompt: &Prompt) {
-    let block = Block::default()
-        .title("Join Lichess Game")
-        .borders(Borders::ALL)
-        .border_type(BorderType::Rounded)
-        .padding(Padding::horizontal(1))
-        .border_style(Style::default().fg(WHITE));
-    let area = centered_rect(50, 30, frame.area());
-
-    let current_input = prompt.input.as_str();
-
-    let text = vec![
-        Line::from("Enter game ID or URL:").alignment(Alignment::Center),
-        Line::from(""),
-        Line::from(current_input),
-        Line::from(""),
-        Line::from(""),
-        Line::from("You can paste the full URL or just the game ID."),
-        Line::from("Note: You must be a participant in the game."),
-        Line::from(""),
-        Line::from("Press `Esc` to cancel.").alignment(Alignment::Center),
-    ];
-
-    let paragraph = Paragraph::new(text)
-        .block(block.clone())
-        .alignment(Alignment::Left)
-        .wrap(Wrap { trim: true });
-
-    frame.set_cursor_position(Position::new(
-        // Draw the cursor at the current position in the input field.
-        // This position is can be controlled via the left and right arrow key
-        area.x + prompt.character_index as u16 + 2,
-        // Move one line down, from the border to the input line
-        area.y + 3,
-    ));
-
-    frame.render_widget(Clear, area); //this clears out the background
-    frame.render_widget(block, area);
-    frame.render_widget(paragraph, area);
-}
-
-// This renders a popup allowing us to enter a Lichess API token
-pub fn render_enter_lichess_token_popup(frame: &mut Frame, prompt: &Prompt) {
-    let block = Block::default()
-        .title("Enter Lichess API Token")
-        .borders(Borders::ALL)
-        .border_type(BorderType::Rounded)
-        .padding(Padding::horizontal(1))
-        .border_style(Style::default().fg(WHITE));
-    let area = centered_rect(70, 40, frame.area());
-
-    let current_input = prompt.input.as_str();
-    // Mask the token for security (show only last 4 characters)
-    let masked_input = if current_input.len() > 4 {
-        format!(
-            "{}...{}",
-            "*".repeat(current_input.len().saturating_sub(4)),
-            &current_input[current_input.len().saturating_sub(4)..]
-        )
-    } else if !current_input.is_empty() {
-        "*".repeat(current_input.len())
-    } else {
-        String::new()
-    };
-
-    let text = vec![
-        Line::from("Enter your Lichess API token:").alignment(Alignment::Center),
-        Line::from(""),
-        Line::from(masked_input),
-        Line::from(""),
-        Line::from(""),
-        Line::from("To get a token, follow the documentation:"),
-        Line::from("Documentation: https://thomas-mauran.github.io/chess-tui/docs/Lichess/setup"),
-        Line::from(""),
-        Line::from("Press `Enter` to save, `Esc` to cancel.").alignment(Alignment::Center),
-    ];
-
-    let paragraph = Paragraph::new(text)
-        .block(block.clone())
-        .alignment(Alignment::Left)
-        .wrap(Wrap { trim: true });
-
-    frame.set_cursor_position(Position::new(
-        // Draw the cursor at the current position in the input field.
-        // This position is can be controlled via the left and right arrow key
-        area.x + prompt.character_index as u16 + 2,
-        // Move one line down, from the border to the input line
-        area.y + 3,
-    ));
-
-    frame.render_widget(Clear, area); //this clears out the background
-    frame.render_widget(block, area);
-    frame.render_widget(paragraph, area);
-}
